fix: update autosummary cross-reference syntax
New versions of autosummary make it much more difficult to specify cross
references.  Before, it was always correct to use fully-qualified names.
Now, if the autosummary directive is contained within a class, its cross
references must not start with the name of the module that that class is
defined in.  In other words, the cross references must be relative, not
absolute, in this case.  If the autosummary directive is not contained
within a class, then fully-qualified names are still required.  This
adds a bunch of annoying complexity to the code.

Fixes #39